+++ components/Buttons/src/ColorThemer/MDCContainedButtonColorThemer.m
@@ -25,9 +25,13 @@
   [button setBackgroundColor:[colorScheme.onSurfaceColor colorWithAlphaComponent:0.12f]
                     forState:UIControlStateDisabled];
   [button setTitleColor:colorScheme.onPrimaryColor forState:UIControlStateNormal];
-  [button setTitleColor:[colorScheme.onSurfaceColor colorWithAlphaComponent:0.26f]
+  [button setTitleColor:[colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f]
                forState:UIControlStateDisabled];
+  [button setImageTintColor:colorScheme.onPrimaryColor forState:UIControlStateNormal];
+  [button setImageTintColor:[colorScheme.onSurfaceColor colorWithAlphaComponent:0.38f]
+                   forState:UIControlStateDisabled];
   button.disabledAlpha = 1.f;
+  button.inkColor = [colorScheme.onPrimaryColor colorWithAlphaComponent:0.32f];
 }
 
 + (void)resetUIControlStatesForButtonTheming:(nonnull MDCButton *)button {
@@ -36,6 +40,7 @@
   for (NSUInteger state = 0; state <= maximumStateValue; ++state) {
     [button setBackgroundColor:nil forState:state];
     [button setTitleColor:nil forState:state];
+    [button setImageTintColor:nil forState:state];
   }
 }