+++ components/Buttons/src/ColorThemer/MDCFloatingButtonColorThemer.h
@@ -14,20 +14,21 @@
  limitations under the License.
  */
 
-#import "MDCTextButtonThemer.h"
+#import "MaterialButtons.h"
+#import "MaterialColorScheme.h"
 
-#import "MDCTextButtonColorThemer.h"
-#import "MDCButtonTypographyThemer.h"
+/**
+ A color themer for a Material floating Button.
+ */
+@interface MDCFloatingButtonColorThemer : NSObject
 
-@implementation MDCTextButtonThemer
+/**
+ Applies a color scheme to theme to an MDCFloatingButton.
 
-+ (void)applyScheme:(nonnull id<MDCButtonScheming>)scheme
-           toButton:(nonnull MDCButton *)button {
-  [MDCTextButtonColorThemer applySemanticColorScheme:scheme.colorScheme toButton:button];
-  [MDCButtonTypographyThemer applyTypographyScheme:scheme.typographyScheme toButton:button];
-  button.minimumSize = CGSizeMake(0, scheme.minimumHeight);
-  button.layer.cornerRadius = scheme.cornerRadius;
-}
+ @param colorScheme The color scheme to apply to @c floatingButton.
+ @param button An MDCFloatingButton instance to apply a color scheme.
+ */
++ (void)applySemanticColorScheme:(nonnull id<MDCColorScheming>)colorScheme
+                        toButton:(nonnull MDCFloatingButton *)button;
 
 @end
-